chore: move metadata to home page
